Add support for datadog prom federation, minor tweaks
The main change of this commit is to support having datadog scrape from
the prometheus federation endpoint. This is useful along with the small
change for adding in custom relabel configs, allow for custom handling
of labels and then getting those modified labels values into datadog.

There is also a small fix here for making the websocket work with a
trusted CA