Run the cleanup in the bookie init job ---
*Motivation*

The pulsar cluster init job will run after the bookie init.
If we restore a cluster with the same name of the old cluster,
it will fail to start bookie with mismatch cookie error.
So we allow to cleanup /ledgers path to cleanup the bookie metadata.
So we can initialize the cluster with the restore point successfully.